+++ tests/conftest.py
@@ -1,5 +1,6 @@
 import collections
 import os
+from textwrap import dedent
 
 import pytest
 import yaml
@@ -44,28 +45,40 @@ def testing_workdir(tmpdir, request):
 @pytest.fixture(scope="function")
 def config_yaml(testing_workdir):
     config = {"python": ["2.7", "3.5"], "r_base": ["3.3.2", "3.4.2"]}
+    os.makedirs(os.path.join(testing_workdir, "recipe"))
+    os.makedirs(os.path.join(testing_workdir, "migrations"))
     with open(os.path.join(testing_workdir, "config.yaml"), "w") as f:
+        f.write("docker:\n")
+        f.write("  fallback_image:\n")
+        f.write("  - centos:6\n")
+    with open(os.path.join(testing_workdir, "recipe", "default_config.yaml"), "w") as f:
         yaml.dump(config, f, default_flow_style=False)
         # need selectors, so write these more manually
-        f.write("target_platform:\n")
-        f.write("- win-64   # [win]\n")
-        f.write("- win-32   # [win]\n")
-        f.write("c_compiler:\n  # [win]")
-        f.write("- vs2008\n  # [win]")
-        f.write("- vs2015\n  # [win]")
-        f.write("zip_keys:\n  # [win]")
-        f.write("- c_compiler\n   # [win]")
-        f.write("- python\n   # [win]")
+        f.write(dedent("""\
+        target_platform: 
+        - win-64        # [win]
+        - win-32        # [win]
+        c_compiler:     # [win]
+        - vs2008        # [win]
+        - vs2015        # [win]
+        vc:             # [win]
+        - '9'           # [win]
+        - '14'          # [win]
+        zip_keys:       # [win]
+        - c_compiler    # [win]
+        - python        # [win]
+        - vc            # [win]
+        """))
     # dummy file that needs to be present for circle ci.  This is created by the init function
     os.makedirs(os.path.join(testing_workdir, ".circleci"))
     with open(
         os.path.join(testing_workdir, ".circleci", "checkout_merge_commit.sh"), "w"
     ) as f:
         f.write("echo dummy file")
-    with open(os.path.join(testing_workdir, "short_config.yaml"), "w") as f:
+    with open(os.path.join(testing_workdir, "recipe", "short_config.yaml"), "w") as f:
         config = {"python": ["2.7"]}
         yaml.dump(config, f, default_flow_style=False)
-    with open(os.path.join(testing_workdir, "long_config.yaml"), "w") as f:
+    with open(os.path.join(testing_workdir, "recipe", "long_config.yaml"), "w") as f:
         config = {"python": ["2.7", "3.5", "3.6"]}
         yaml.dump(config, f, default_flow_style=False)
     return testing_workdir
@@ -73,7 +86,6 @@ def config_yaml(testing_workdir):
 
 @pytest.fixture(scope="function")
 def noarch_recipe(config_yaml, request):
-    os.makedirs(os.path.join(config_yaml, "recipe"))
     with open(os.path.join(config_yaml, "recipe", "meta.yaml"), "w") as fh:
         fh.write(
             """
@@ -92,14 +104,14 @@ requirements:
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(config_yaml, "config.yaml")
+            config_yaml, exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml")
         ),
     )
 
 
 @pytest.fixture(scope="function")
 def r_recipe(config_yaml, request):
-    os.makedirs(os.path.join(config_yaml, "recipe"))
     with open(os.path.join(config_yaml, "recipe", "meta.yaml"), "w") as fh:
         fh.write(
             """
@@ -118,14 +130,14 @@ requirements:
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(config_yaml, "config.yaml")
+            config_yaml, exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml")
         ),
     )
 
 
 @pytest.fixture(scope="function")
 def py_recipe(config_yaml, request):
-    os.makedirs(os.path.join(config_yaml, "recipe"))
     with open(os.path.join(config_yaml, "recipe", "meta.yaml"), "w") as fh:
         fh.write(
             """
@@ -146,14 +158,97 @@ about:
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(config_yaml, "config.yaml")
+            config_yaml, exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml")
+        ),
+    )
+
+
+@pytest.fixture(scope="function")
+def recipe_migration_cfep9(config_yaml, request):
+    # write a migrator
+    with open(os.path.join(config_yaml, "recipe", "meta.yaml"), "w") as fh:
+        fh.write(
+            """
+package:
+    name: py-test
+    version: 1.0.0
+requirements:
+    host:
+        - python
+        - zlib
+    run:
+        - python
+about:
+    home: home
+    """
+        )
+
+    with open(os.path.join(config_yaml, "migrations", "zlib.yaml"), "w") as fh:
+        fh.write("""
+zlib:
+    - 1000
+""")
+
+    return RecipeConfigPair(
+        str(config_yaml),
+        _load_forge_config(
+            config_yaml, exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml")
+        ),
+    )
+
+
+@pytest.fixture(scope="function")
+def recipe_migration_cfep9_downgrade(config_yaml, recipe_migration_cfep9):
+    # write a downgrade migrator that lives next to the current migrator.
+    # Only this, more recent migrator should apply.
+    with open(os.path.join(config_yaml, "migrations", "zlib-downgrade.yaml"), "w") as fh:
+        fh.write("""
+migration_ts: 1.0
+zlib:
+    - 999
+""")
+    #return recipe_migration_cfep9
+    return RecipeConfigPair(
+        str(config_yaml),
+        _load_forge_config(
+            config_yaml, exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml")
+        ),
+    )
+
+@pytest.fixture(scope="function")
+def recipe_migration_win_compiled(config_yaml, py_recipe):
+    with open(os.path.join(config_yaml, "migrations", "vc-migrate.yaml"), "w") as fh:
+        fh.write(dedent("""
+        migration_ts: 1.0
+        c_compiler:    # [win]
+            - vs2008   # [win]
+            - vs2017   # [win]
+        cxx_compiler:  # [win]
+            - vs2008   # [win]
+            - vs2017   # [win]
+        vc:            # [win]
+            - '9'      # [win]
+            - '14.1'   # [win]
+        zip_keys:
+            - - python          # [win]
+              - c_compiler      # [win]
+              - cxx_compiler    # [win]
+              - vc              # [win]
+        """))
+    return RecipeConfigPair(
+        str(config_yaml),
+        _load_forge_config(
+            config_yaml, exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml")
         ),
     )
 
 
 @pytest.fixture(scope="function")
 def skipped_recipe(config_yaml, request):
-    os.makedirs(os.path.join(config_yaml, "recipe"))
     with open(os.path.join(config_yaml, "recipe", "meta.yaml"), "w") as fh:
         fh.write(
             """
@@ -174,14 +269,14 @@ about:
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(config_yaml, "config.yaml")
+            config_yaml, exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml")
         ),
     )
 
 
 @pytest.fixture(scope="function")
 def python_skipped_recipe(config_yaml, request):
-    os.makedirs(os.path.join(config_yaml, "recipe"))
     with open(os.path.join(config_yaml, "recipe", "meta.yaml"), "w") as fh:
         fh.write(
             """
@@ -202,14 +297,14 @@ about:
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(config_yaml, "config.yaml")
+            config_yaml, exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml")
         ),
     )
 
 
 @pytest.fixture(scope="function")
 def linux_skipped_recipe(config_yaml, request):
-    os.makedirs(os.path.join(config_yaml, "recipe"))
     with open(os.path.join(config_yaml, "recipe", "meta.yaml"), "w") as fh:
         fh.write(
             """
@@ -228,7 +323,8 @@ about:
     return RecipeConfigPair(
         str(config_yaml),
         _load_forge_config(
-            config_yaml, exclusive_config_file=os.path.join(config_yaml, "config.yaml")
+            config_yaml, exclusive_config_file=os.path.join(
+                config_yaml, "recipe", "default_config.yaml")
         ),
     )