+++ tests/test_lint_recipe.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-from __future__ import print_function, unicode_literals
 from collections import OrderedDict
 from contextlib import contextmanager
 import io
@@ -11,7 +10,9 @@ import tempfile
 import textwrap
 import unittest
 import warnings
+
 import github
+import pytest
 
 import nwb_extensions_smithy.lint_recipe as linter
 
@@ -568,6 +569,20 @@ class Test_linter(unittest.TestCase):
         )
         self.assertIn(expected_message, lints)
 
+    def test_license_file_required(self):
+        meta = {
+            "about": {
+                "home": "a URL",
+                "summary": "A test summary",
+                "license": "MIT",
+            }
+        }
+        lints, hints = linter.lintify(meta)
+        expected_message = (
+            "license_file entry is missing, but is required."
+        )
+        self.assertIn(expected_message, lints)
+
     def test_recipe_name(self):
         meta = {"package": {"name": "mp++"}}
         lints, hints = linter.lintify(meta)
@@ -803,6 +818,7 @@ class Test_linter(unittest.TestCase):
         self.assertIn(msg, lints)
 
 
+@pytest.mark.cli
 class TestCLI_recipe_lint(unittest.TestCase):
     def test_cli_fail(self):
         with tmp_directory() as recipe_dir: